+++ AirCasting/CoreData/ExternalSessionEntity+CoreDataProperties.swift
@@ -4,7 +4,6 @@
 
 import Foundation
 import CoreData
-import CoreLocation
 
 extension ExternalSessionEntity {
     
@@ -107,42 +106,5 @@ extension ExternalSessionEntity {
     
     private func streamWithSensorNameContaining(_ parameterName: String) -> MeasurementStreamEntity? {
         allStreams.first(where: { $0.sensorName?.contains(parameterName) ?? false })
-    }
-}
-
-extension ExternalSessionEntity: Sessionable {
-    var gotDeleted: Bool {
-        false
-    }
-    
-    var userInterface: UIStateEntity? {
-        get { uiState }
-        set { uiState = newValue }
-    }
-    
-    var location: CLLocationCoordinate2D? {
-        get {
-            guard let lat = value(forKey: "latitude") as? CLLocationDegrees,
-                  let lon = value(forKey: "longitude") as? CLLocationDegrees else {
-                      return nil
-                  }
-            return CLLocationCoordinate2D(latitude: lat, longitude: lon)
-        }
-        set {
-            setValue(newValue?.latitude, forKey: "latitude")
-            setValue(newValue?.longitude, forKey: "longitude")
-        }
-    }
-    
-    var isFixed: Bool {
-        return true
-    }
-    
-    var isExternal: Bool {
-        return true
-    }
-    
-    var isActive: Bool {
-        return true
     }
 }
