+++ AirCasting/CoreData/SessionEntity+Sessionable.swift
@@ -1,0 +1,8 @@
+// Created by Lunar on 24/05/2022.
+//
+
+import Foundation
+
+extension SessionEntity: Sessionable {
+    var isExternal: Bool { false }
+}
